+++ jwst/refpix/refpix_step.py
@@ -25,11 +25,11 @@ class RefPixStep(Step):
         ovr_corr_mitigation_ftr = float(default=3.0) # Factor to avoid overcorrection of bad reference pixels for IRS2
         preserve_irs2_refpix = boolean(default=False) # Preserve reference pixels in output
         irs2_mean_subtraction = boolean(default=False) # Apply a mean offset subtraction before IRS2 correction
-        use_conv_kernel = boolean(default=True) # For NIR data only, use the convolution kernel instead of the running median
-        sigreject = integer(default=4) # Number of sigmas to reject as outliers
-        gaussmooth = integer(default=1) # Width of Gaussian smoothing kernel to use as a low-pass filter
-        halfwidth = integer(default=30) # Half-width of convolution kernel to build
-        user_supplied_reffile = string(default=None)  # ASDF user-supplied reference file
+        use_conv_kernel = boolean(default=True) # For NIR full-frame data, use convolution kernel instead of running median
+        sigreject = float(default=4.0) # Number of sigmas to reject as outliers
+        gaussmooth = float(default=1.0) # Width of Gaussian smoothing kernel to use as a low-pass filter
+        halfwidth = float(default=30.0) # Half-width of convolution kernel to build
+        user_supplied_reffile = string(default=None)  # ASDF user-supplied reference file for convolution kernel
     """
 
     reference_file_types = ['refpix']
